ファイルの分け方変更
